More mpi problems
